aws: Adapt aws provider code to use the common provider framework
Fixes:#579
