+++ pkg/adaptor/cloud/aws/provider.go
@@ -1,0 +1,171 @@
+// (C) Copyright Confidential Containers Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+package aws
+
+import (
+	"context"
+	"encoding/base64"
+	"errors"
+	"fmt"
+	"log"
+	"net"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/ec2"
+	"github.com/aws/aws-sdk-go-v2/service/ec2/types"
+
+	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor/cloud"
+	"github.com/confidential-containers/cloud-api-adaptor/pkg/util"
+	"github.com/confidential-containers/cloud-api-adaptor/pkg/util/cloudinit"
+)
+
+var logger = log.New(log.Writer(), "[adaptor/cloud/aws] ", log.LstdFlags|log.Lmsgprefix)
+var errNotReady = errors.New("address not ready")
+
+const maxInstanceNameLen = 63
+
+type awsProvider struct {
+	ec2Client     *ec2.Client
+	serviceConfig *Config
+}
+
+func NewProvider(config *Config) (cloud.Provider, error) {
+
+	logger.Printf("aws config: %#v", config.Redact())
+
+	if err := retrieveMissingConfig(config); err != nil {
+		logger.Printf("Failed to retrieve configuration, some fields may still be missing: %v", err)
+	}
+
+	ec2Client, err := NewEC2Client(*config)
+	if err != nil {
+		return nil, err
+	}
+
+	provider := &awsProvider{
+		ec2Client:     ec2Client,
+		serviceConfig: config,
+	}
+
+	return provider, nil
+}
+
+func getIPs(instance types.Instance) ([]net.IP, error) {
+
+	var podNodeIPs []net.IP
+	for i, nic := range instance.NetworkInterfaces {
+		addr := nic.PrivateIpAddress
+
+		if addr == nil || *addr == "" || *addr == "0.0.0.0" {
+			return nil, errNotReady
+		}
+
+		ip := net.ParseIP(*addr)
+		if ip == nil {
+			return nil, fmt.Errorf("failed to parse pod node IP %q", *addr)
+		}
+		podNodeIPs = append(podNodeIPs, ip)
+
+		logger.Printf("podNodeIP[%d]=%s", i, ip.String())
+	}
+
+	return podNodeIPs, nil
+}
+
+func (p *awsProvider) CreateInstance(ctx context.Context, podName, sandboxID string, cloudConfig cloudinit.CloudConfigGenerator) (*cloud.Instance, error) {
+
+	instanceName := util.GenerateInstanceName(podName, sandboxID, maxInstanceNameLen)
+
+	userData, err := cloudConfig.Generate()
+	if err != nil {
+		return nil, err
+	}
+
+	//Convert userData to base64
+	userDataEnc := base64.StdEncoding.EncodeToString([]byte(userData))
+
+	var input *ec2.RunInstancesInput
+
+	if p.serviceConfig.UseLaunchTemplate {
+		input = &ec2.RunInstancesInput{
+			MinCount: aws.Int32(1),
+			MaxCount: aws.Int32(1),
+			LaunchTemplate: &types.LaunchTemplateSpecification{
+				LaunchTemplateName: aws.String(p.serviceConfig.LaunchTemplateName),
+			},
+			UserData: &userDataEnc,
+		}
+	} else {
+		input = &ec2.RunInstancesInput{
+			MinCount:         aws.Int32(1),
+			MaxCount:         aws.Int32(1),
+			ImageId:          aws.String(p.serviceConfig.ImageId),
+			InstanceType:     types.InstanceType(p.serviceConfig.InstanceType),
+			SecurityGroupIds: p.serviceConfig.SecurityGroupIds,
+			SubnetId:         aws.String(p.serviceConfig.SubnetId),
+			UserData:         &userDataEnc,
+		}
+		if p.serviceConfig.KeyName != "" {
+			input.KeyName = aws.String(p.serviceConfig.KeyName)
+		}
+	}
+
+	logger.Printf("CreateInstance: name: %q", instanceName)
+
+	result, err := p.ec2Client.RunInstances(ctx, input)
+	if err != nil {
+		return nil, fmt.Errorf("Creating instance (%v) returned error: %s", result, err)
+	}
+
+	logger.Printf("created an instance %s for sandbox %s", *result.Instances[0].PublicDnsName, sandboxID)
+
+	tagInput := &ec2.CreateTagsInput{
+		Resources: []string{*result.Instances[0].InstanceId},
+		Tags: []types.Tag{
+			{
+				Key:   aws.String("Name"),
+				Value: aws.String(instanceName),
+			},
+		},
+	}
+
+	_, err = p.ec2Client.CreateTags(ctx, tagInput)
+	if err != nil {
+		logger.Printf("Adding tags to the instance failed with error: %s", err)
+	}
+
+	instanceID := *result.Instances[0].InstanceId
+
+	ips, err := getIPs(result.Instances[0])
+	if err != nil {
+		logger.Printf("failed to get IPs for the instance : %v ", err)
+		return nil, err
+	}
+
+	instance := &cloud.Instance{
+		ID:   instanceID,
+		Name: instanceName,
+		IPs:  ips,
+	}
+
+	return instance, nil
+}
+
+func (p *awsProvider) DeleteInstance(ctx context.Context, instanceID string) error {
+	terminateInput := &ec2.TerminateInstancesInput{
+		InstanceIds: []string{
+			instanceID,
+		},
+	}
+
+	resp, err := p.ec2Client.TerminateInstances(ctx, terminateInput)
+
+	if err != nil {
+		logger.Printf("failed to delete an instance: %v and the response is %v", err, resp)
+		return err
+	}
+	logger.Printf("deleted an instance %s", instanceID)
+	return nil
+
+}
