+++ pkg/adaptor/cloud/aws/types_test.go
@@ -1,4 +1,5 @@
-//go:build aws
+// (C) Copyright Confidential Containers Contributors
+// SPDX-License-Identifier: Apache-2.0
 
 package aws
