+++ pkg/adaptor/hypervisor/registry/aws.go
@@ -1,13 +1,12 @@
-// +build aws
+//go:build aws
 
 package registry
 
 import (
 	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor/hypervisor"
-	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor/hypervisor/aws"
 	"github.com/confidential-containers/cloud-api-adaptor/pkg/podnetwork"
 )
 
 func newServer(cfg hypervisor.Config, cloudConfig interface{}, workerNode podnetwork.WorkerNode, daemonPort string) hypervisor.Server {
-	return aws.NewServer(cfg, cloudConfig.(aws.Config), workerNode, daemonPort)
+	panic("never reaches here. this code will be removed when refactoring is done")
 }
